requirements.txt
Django==1.11.8
psycopg2==2.7.3.2
djangorestframework==3.7.3
django-tables2==1.17.1
django-filter==1.1.0
django-email-obfuscator==0.1.5
wget==3.2
requests==2.18.4
beautifulsoup4==4.6.0
django-crispy-forms==1.7.0
tablib==0.12.1
